Only apply IR patches for jflte

+++ roms/cyanogenmod.sh
@@ -52,30 +52,32 @@ cyanogenmod_prebuild() {
   # Remove old zips
   rm -f "${OUT}"/cm-*.zip*
 
-  pushd frameworks/base/
-  git checkout github/${REPO_BRANCH}
-  git clean -fdx
-  git diff | patch -p1 -R
-  #git fetch http://review.cyanogenmod.org/CyanogenMod/android_frameworks_base refs/changes/80/44580/9 && git cherry-pick FETCH_HEAD
-  git am ${WORKSPACE}/hudson/roms/${REPO_BRANCH}/0001-Irda-Add-Irda-System-Service-for-Samsung-devices.patch
-  popd
+  if grep -q jflte <<< ${LUNCH}; then
+    pushd frameworks/base/
+    git checkout github/${REPO_BRANCH}
+    git clean -fdx
+    git diff | patch -p1 -R
+    #git fetch http://review.cyanogenmod.org/CyanogenMod/android_frameworks_base refs/changes/80/44580/9 && git cherry-pick FETCH_HEAD
+    git am ${WORKSPACE}/hudson/roms/${REPO_BRANCH}/0001-Irda-Add-Irda-System-Service-for-Samsung-devices.patch
+    popd
 
-  pushd hardware/libhardware/
-  git checkout github/${REPO_BRANCH}
-  git clean -fdx
-  git diff | patch -p1 -R
-  #git fetch http://review.cyanogenmod.org/CyanogenMod/android_hardware_libhardware refs/changes/83/44783/4 && git checkout FETCH_HEAD
-  rm -f include/hardware/irda.h
-  git am ${WORKSPACE}/hudson/roms/${REPO_BRANCH}/0001-Irda-Added-IrDA-HAL-Library.patch
-  popd
+    pushd hardware/libhardware/
+    git checkout github/${REPO_BRANCH}
+    git clean -fdx
+    git diff | patch -p1 -R
+    #git fetch http://review.cyanogenmod.org/CyanogenMod/android_hardware_libhardware refs/changes/83/44783/4 && git checkout FETCH_HEAD
+    rm -f include/hardware/irda.h
+    git am ${WORKSPACE}/hudson/roms/${REPO_BRANCH}/0001-Irda-Added-IrDA-HAL-Library.patch
+    popd
 
-  pushd device/samsung/jf-common/
-  git checkout github/${REPO_BRANCH}
-  git clean -fdx
-  git diff | patch -p1 -R
-  #git fetch http://review.cyanogenmod.org/CyanogenMod/android_device_samsung_jf-common refs/changes/91/44691/5 && git checkout FETCH_HEAD
-  git am ${WORKSPACE}/hudson/roms/${REPO_BRANCH}/0001-Irda-Enable-Irda-service-via-overlay-and-HAL.patch
-  popd
+    pushd device/samsung/jf-common/
+    git checkout github/${REPO_BRANCH}
+    git clean -fdx
+    git diff | patch -p1 -R
+    #git fetch http://review.cyanogenmod.org/CyanogenMod/android_device_samsung_jf-common refs/changes/91/44691/5 && git checkout FETCH_HEAD
+    git am ${WORKSPACE}/hudson/roms/${REPO_BRANCH}/0001-Irda-Enable-Irda-service-via-overlay-and-HAL.patch
+    popd
+  fi
 
   if [ ! -z "${CM_NIGHTLY}" ]; then
     make update-api
